NavigationActionBar: don't override the ToolBar background
We did this so that we could hide the separator and dim the background
on mobile mode. Actually we can just leave the background style as-is
and instead set the background to be invisible.

Also merge HeaderFooterToolbar into NavigationActionBar since without
the custom styling it is basically just a regular ToolBar.