MediaPlayListProxyModel: only enqueue after the current track when shuffled
and only after the current album when in album shuffle mode.

Typically if a user wants to enqueue a track, they also want to hear the
track at some point.

Also fixes an issue where the enqueued track could sometimes enqueue at
the index of the currently playing track, causing the current track to
change.

BUG: 488690