Use implicit height for main toolbar rather than setting an explicit height
Previously we were fixing the height of the main toolbar, which was smaller
than the button size of some styles (E.g. Universal Qt Quick style). Now
the toolbar size increases to fit the content. The minimum toolbar size
is unchanged, so there is no visual change when running with qqc2-desktop-style
with Breeze QWidgets style.