Reduce choppyness of animation when toggling playlist visibility
Changing the playlist's width causes many Loaders to activate/deactivate,
which results in a choppy animation when hiding the playlist. When hiding
the playlist, we fix the PlayList width and change the parent item's width
instead. This gives the visual of the PlayList sliding off the screen without
the performance issues caused by Loaders.