Remove docbook
This app is fairly simple and easy to use, and in general people don't
read docs for these kinds of apps. In addition, Elisa's docs are
woefully out of date, currently of no use to anyone.

Instead of putting effort into improving them, let's instead delete them
and put that effort into improving the UI so anything that might need to
be explained by docs no one will read is instead obvious in the UI, as
the HIG recommends.