Don't use gridUnit as an icon size
This was not correct in the past when it reflected the default text
height, nor in the present where it's hardcoded to 18px; in both cases,
it'll render the icon in a blurry way since we don't use 18px for icons.

Instead, let the buttons have their default icon sizes coming from the
theme.