Save and restore maximization state
Right now Elisa does not save and restore maximization state, and does
something annoying instead: it makes the unmaximized window have the
width and height of a maximized window!

Fix this by tracking maximization state explicitly. It needs to be done
in a new onClosing signal handler because by the time the current
aboutToQuit() binding executes, the window state has been changed from
maximized to windowed internally by Qt.

For 24.12, we'll be able to migrate to KConfig.WindowStateSaver,
assuming its own maximization state tracking gets fixed. For now, add
the feature to the existing implementation.

BUG: 428510
FIXED-IN: 24.08.0