use tabular numerals for durations and positions
Proportional width fonts that don't default to tabular numerals will
lead to broken lines for tabular display of numbers like durations and
playlist positions, as well as jumping minute numbers for current play
positions. This can also shift the alignment of dependent controls like
buttons. An example for such a font is Inter v4.0.

This change explicitly requests tabular numerals for durations in the
play controls, the playlist, the track browser, and the remaining
playlist duration in the footer, and for playlist positions as well as
the currently playing position. The base metrics are set to use them as
well.

BUG: 491714