HeaderBar: unmaximize when track information clicked
When the track information in the HeaderBar is clicked, the main display
switches accordingly, to the Now Playing view, the Artist view, or the
Album view. This also works in Party mode (i.e. when the  HeaderBar is
maximized); the view changes in just the same way. But as the state
stays maximized, it appears like nothing is happening. Given that this
click signals a clear user intention to access this functionality, it
makes sense to make it immediately accessible without requiring the user
to leave Party mode manually.

This change unmaximizes the HeaderBar when the user clicks one of these
three lines.