make main drawer handle use an appropriate color
The ContextDrawer that contains the playlist has its handle show up in
two circumstances: in Party mode, and for narrow windows where there is
not enough space to display the playlist directly (in desktop mode). It
is always displayed as part of the HeaderBar area background that is
colored using a blurred version of the album art; as this always leads
to a dark-ish image, the labels and controls there need to use a
suitable color. For the labels (e.g. track title) this is hardcoded, but
not for the drawer handle, which takes the color from the active theme.
For light color schemes like Breeze light, this leads to a barely
visible black-on-dark icon.

This change moves the hardcoded color value to the BaseTheme (as it is
now accessed in multiple places) and sets the color of the handle to
the same value, leading to a visible icon and consistent colors in the
HeaderBar.