ElisaMainWindow: set drawer handle tooltips
The handle for opening and closing the ContextDrawer currently does not
set a tooltip explicitly, inheriting the generic Kirigami value of
"Open/Close drawer". This is not very informative.

This change sets the tooltip to "Show/Hide playlist drawer". The word
drawer is not omitted, as the playlist may already be visible, namely in
Party mode which contains a large, but non-editable playlist, and the
extra word helps disambiguate the two.