MPC: use sentence case for tooltips & de-"Toggle"
The HIG recommend that tooltips use sentence case. The tooltip-having
buttons in MediaPlayerControl are all icon-only (except "Show Playlist")
and the text is only used as a tooltip, but use titlecase anyway as if
they were button labels. In addition, some like party mode and mute use
a "Toggle X" construction whether on or off, while such tooltips usually
show the action that the button does when clicked (e.g. in the Plasma
volume applet, "Mute" or "Unmute", rather than "Toggle Mute").

This change uses sentence case consistently for labels that are shown as
tooltips (leaving "Show Playlist" as is), adds information that this
will be displayed as a tooltip to the translation context for languages
that have similar rules as English, and changes the tooltips away from a
"Toggle X" construction.