NAB: use sentence case for tooltips & minor fixes
The HIG recommend that tooltips use sentence case. The buttons in
NavigationActionBar, similarly to MediaPlayerControl in the previous
commit, mostly use Title case, but the text is mostly displayed as
tooltips. In this case, however, the buttons may also be icon-and-text
on mobile in landscape orientation, and so need to have both texts
explicitly set. There are also a small number of wording issues, such as
using "Create a Radio" as the tool tip for the Add button, which could
be read as being a button that begins a new internet stream for others
to listen to, rather than adding an existing one to the list of stations
available in Elisa.

This change explicitly sets sentence case tooltips in
NavigationActionBar and implements some small wording fixes, including
one case of de-"toggle"-ing, labeling the button "Add radio station"
instead of "Create a Radio". To make the new tooltips work, the icon
for changing view mode now show the state that will be switched into,
rather than the state that the view is in, which is already clear
just by looking at the view.