SettingsForm: follow HIG on capitalization
The HIG recommend the use of sentence case for checkbox labels and many
other UI elements. There are a few instances where the Settings dialog
does not follow this, such as "Scan for New Music". A few more cases
involve things that might appear to be proper names but are probably
better treated as generic names, as they need not match the current or
future names in Plasma (e.g. there is no "Task Manager" by itself, only
the "Icons-only Task Manager" and its Icons-and-Text sibling.

This change makes the use of sentence case consistent; only page names
in Elisa itself (e.g. "Now Playing") are kept capitalized. In addition,
the folder picker dialogs are now titled "Choose Folder" rather than
"Choose a Folder", to be more consistent with the guidance on omitting
uneccesary (in the UI context) articles.