HeaderBar: separate Exit Full Screen and handle
In party mode, Elisa shows a ContextDrawer handle in the top right
corner that gives access to playlist controls. In full screen mode,
Elisa shows an Exit Full Screen button in the top right corner. If party
mode is in full screen, both buttons are in the top right corner and
take the same space. In addition, they are different kinds of buttons -
the handle is hardcoded as a ToolButton on the Kirigami side, the Exit
Full Screen is a regular button to maintain visibility with all themes
on the dark-ish background of the header bar.

This change separates the two buttons, with the drawer handle on the
corner (as it pops out something from the side of the screen). The Exit
Full Screen button is also converted to a flat button, with text and
icon colors set to the HeaderBar text color, which the handle also uses,
to maintain visibility with all themes.