SimplePlayListView: disable animations
Qt ListView animations and section headers currently do not play well
together, see https://bugreports.qt.io/browse/QTBUG-49868. For this
reason, animations in the MediaPlayListView were temporarily disabled
years ago (see Elisa MR !68). This was not done in SimplePlayListView;
it's less common to have animated changes here (as it usually involves
using the playlist sidebar or adding files to the playlist using another
program), but when it happens there are graphical glitches just like
there were in the main view before they were disabled.

This change comments out the animations (the same way this was handled
in MediaPlayListView), with a link to the Qt bug and the original MR for
further information.