SimplePlayListView: remove hardcoded text color
SimplePlayListView currently explicitly hardcodes the text color for
album headers to that of Breeze Dark. This is done presumably because
this will appear in the HeaderBar playlist in party mode, and the text
needs to be readable on dark background. It's not necessary though, as
HeaderBar itself sets the text color already and SimplePlayListView
inherits this; the playlist itself does not set a text color and still
is colored appropriately. Currently this is the only use of this
component, but if it gets used elsewhere, there's no reason to assume
that text color of only the album header should need to be different.

This change removes the hardcoded text color.