HeaderBar: toggle Party mode with double click
Party mode is currently only accessible with the mouse by hitting a
small target near a window edge. This makes it somewhat cumbersome to
switch between modes, which is a shame because Party mode has a nice big
display for album art and most important controls, but does not afford
adding to or creating new lists.

This change sets a MouseArea in the Header bar that captures double
clicks and toggles Party mode.