HeaderBar: set background image background
The HeaderBar uses a blur and lightness transformation applied to the
cover image as its background, which yields a dark-ish image suitable
for dark-theme foreground colors while being visually appealing and
reminiscent of the album cover. It currently does not take transparent
parts of the image into account, which are unaffected by the lightness
transformation and directly show the (usually hidden) background behind
the background, which is a solid color determined by the loaded theme.
With light themes, this can cause white text to be displayed on a white
background, which is not readable.

This change explicitly puts a Kirigami.Theme.backgroundColor rectangle
behind the image, before subjecting it to the transformation, so proper
lightness reduction is performed, yielding an effect similar to that for
a mostly white album cover on light themes.

BUG: 464865