qml: handle return/enter directly in menu buttons
The menu buttons (SortMenu in NavigationActionBar and ApplicationMenu in
MediaPlayerControl currently activate on mousedown, but also activate on
clicked to handle keyboard activation with return or enter, as the
button sends this event. However the two conflict - sometimes clicking
to open immediately closes the menu, and when the menu is open, it can't
be closed by pressing the button again (unless, as in ApplicationMenu,
the menu is set to modal), as completing the click undoes the work that
the pressed action did.

This change removes the onClicked signal handler for these buttons, and
instead listens directly for the keyboard events, so that the menu
consistently stays open when clicked, can be closed by clicking again,
and still opens when pressing return or enter while the button has
keyboard focus.