HeaderBar: set correct Exit Full Screen button type
The change in 519e438f incorrectly set the display property to an
undefined value. While the intended one is also the default (for the
underlying component, not the one that is instantiated) and thus
there are no UI consequences to this, there is an error reported to the
console.

This change sets the display type to the actually existing
TextBesideIcon instead of the undefined IconAndText.