NavigationActionBar: fix placeholder for singleAlbum
### Reason for the change

Elisa's filter bar has a placeholder that offers the user to search for
an "album name, artist, etc.". This is somewhat misleading in single
album view, as you can't search for a different album than the one
that's already being displayed, and searching for the that album will
never do anything.

This change replaces "album name" in the placeholder with "track name",
as this can be sensible with albums containing many tracks. "artist" is
kept, as it stil makes sense on compilation or split albums.

### Test plan

Enter album view, click on an album. Verify that the placeholder does not refer to searching for an album. Enter track view, verify that searching for an album is suggested.

### Screenshots or screen recordings

| before | after  |
| ------ | ------ |
| ![elisa_album_before](/uploads/9cbf9de78e51f83ae74bc047f639974f/elisa_album_before.png) | ![elisa_album_after](/uploads/0f25edda59f9d6a67f19ab09fb12546a/elisa_album_after.png) |