Fix not showing metadata of webradio stream
Currently when playing a webradio stream with VLC backend, metadata like station name, and artist - title are not shown.

This is due to an unnecessary if-statement restricting changes only for MediaPlayList::ColumnsRoles::IsValidRole, CountRole and IsPlayingRole. The if statement can be safely removed because the following switch-case has a sensible default case.

CCBUG: 467265