Added suppression of metadata update in case of VLC issue 21057
There's an issue in libVLC with metadata of ICYcasts when using SSL/TLS
(https). This results in the last subpath of the stream URL being set as metadata.
This commit checks if the title is the last part of the URL and
suppresses the metadata update in Elisa.

See https://code.videolan.org/videolan/vlc/-/issues/21057