Lyrics: load sidecar .lrc file as song lyrics
It's common to have lyrics in a sidecar file (`.lrc`). Currently Elisa only loads lyrics lived in the ID3 tag and ignores the sidecar file.

This commit adds support for reading from lrc files, but disables the ability to modify the lyrics of an lrc file. TrackDataType now stores the location of the lyrics file, which should make it easy to implement this in the future.

Tested manually:

1. Create an audio file called `sample_track.mp3` and a valid lyrics file `sample_track.lrc` in the same directory
2. Use Elisa to play `sample_track.mp3`
3. Open the Now Playing view
4. Open the file metadata dialog

The lyrics are visible in the metadata dialog and Now Playing view. The lyrics are also correctly synchronized.

BUG: 465716