Fix the spacebar play/pause action when specific buttons have focus
An event filter was applied to some specific buttons with the intention
of overriding the space-bar for the play/pause action, but actually it
had the effect of making the space-bar do nothing when pressed while one
of these buttons had focus.

The space-bar shortcut is already handled by the Kirigami.Action in
ElisaMainWindow, making this event filter unnecessary anyway.

BUG: 476532