qml/ContextView: don't show Lyrics when they're invisible
ContextView attempts to hide the lyrics when not selected or in wide
mode by setting the containing Item's implicitWidth to zero. But this
doesn't work (anymore?); the lyrics are still displayed, centered around
the zero-width position where they would appear (so half off-screen) and
overlapping the metadata area. To hide text this way, clip must be set
to true.

Instead, we might as well directly set visible; this seems clearer than
an implicitWidth + clip trick and is likely faster as well as no
separate checks are necessary about what would fit (which is nothing
anyway).

BUG: 504198
FIXED-IN: 25.04.2