Fix keyboard navigation in playlist entry
Using the left and right arrow buttons to navigate between buttons on
playlist delegates was broken by me in a previous commit. This was
because I swapped out a bunch of manually instantiated buttons for a
Row, without realising that the buttons had their KeyNavigation
properties set.

I can't see any way of getting Row to handles this automatically for us,
so the solution is to just go back to manually instantiating the
buttons.

Test-plan:
- Set focus to the "load playlist" button
- Press tab once to focus the top playlist item
- Press right arrow key multiple times and ensure that each button gets
  focus
- Press left arrow key multiple times and ensure that each button gets
  focus until the focus returns to the delegate

Amends 6acac15793cf8ef2b2a1916767910e87c50b798c