Promote all database error messages from Debug to Critical
A failed query or failure to initialize the database is a pretty big
issue! Promoting to critical makes it easier to debug, as we don't have
to filter out the massive amount of messages that come from enabling the
debug log level.

Also add some missing `Q_EMIT databaseError()` lines.