Fix sidebar icon colour on qqc2-desktop-style
Qqc2-desktop-style implements ItemDelegate using a Kirigami.Icon. The
Icon object has a `selected` property that recolours the icon to the
theme's "highlighted" colour. This property only works for theme icons,
i.e. passed via icon.name (it wouldn't make sense to recolour an
arbitrary image).

Currently the view model passes all icons as url's, presumably because
it's easier to store just a single url rather than a string and a url.
The proper solution would be to add a new field to ViewParameters to
store the icon name, but the way this class is set up makes that quite
painful to do. So for now just apply a hack to remove the prefix part of
the url of theme icons passed by the view model.