Fix artist taphandler being too long in the view toolbar
Previously the taphandler stretched the width of the available area of
the artist label, which was longer than the label itself. This meant the
user could click on an empty space and the artist page would open. Now
we limit the width of the clickable area to the actual width of the
label.

Also combine the TapHandler and HoverArea into one MouseArea.