Let the user open music files from within the app
You can already do this from an external app, from the command line,
and via-drag-and-drop. But we're missing a way to open a music file
from within the app itself.

Add it to the existing "Load" button, which becomes an "Open" button
and can now accept music files as well as playlist files.

FEATURE: 511891
FIXED-IN: 25.12.0