make C++ plugin available to qmllint
qmllint expects the generated module information to be in a directory
corresponding to the module name; in this case "org/kde/elisa". Without
this, everything in a qml file that accesses a type on the module will
cause a qmllint warning. But that's not where add_module puts it by
default for a module that is compiled into the application, we need to
set the output directory manually. (See e.g. Arianna, which does the
same).