qml/ContextView: allow header label to elide
The header label "Now Playing" is not allowed to elide. In English, this
fits even at the smallest window size, but in other languages that may
not be the case, and the header may appear quite broken. The header
label also accessible in the sidebar, so it makes sense to elide this
rather than the button labels.